ARM: DRA7: Add io recalibration delay sequence
If changing to AVS0 voltage is required for development purpose,
there will be some IO timing error versus datasheet. The below
sequence is required to recalibrate the IOs after AVS is done.

The complete sequence is documented in AM572x/DRA TRM SPRUHZ6 (October
2014) Section 18.4.6.1.6 IO Delay Recalibration (page 4006)

http://www.ti.com/lit/ug/spruhz6/spruhz6.pdf#page=4006&zoom=auto,0,350.1

NOTE: This is just part 1 of the complete sequence necessary to handle
IO Delay requirements for DRA7x/AM57x processor family. Steps 6 and 7
(programming the muxes within the isolation region) is to be done.

[nm@ti.com: just some pretty updates]
Signed-off-by: Nishanth Menon <nm@ti.com>
Signed-off-by: Sricharan R <r.sricharan@ti.com>