ARM: OMAP: Change set_pl310_ctrl_reg to be generic
set_pl310_ctrl_reg does use the Secure Monitor Call (SMC) to setup
PL310 control register, however, that is something that is generic
enough to be used for OMAP5 generation of processors as well. The only
difference being the service being invoked for the function.

So, convert the service to a macro and use a generic name (same as
that used in Linux for some consistency). While at that, also add a
data barrier which is necessary as per recommendation.

While at this, switch over to smc #0 instead of handcoded assembly.
To ensure gcc compatibility, steal the strategy used by Linux kernel
for sec extension builds (NOTE: we no longer use '-march=armv5' as the
legacy comment claims).

Signed-off-by: Nishanth Menon <nm@ti.com>